Minor modifications
Change some command line arguments.
Modify logging logic. Wrap the training statistics dictionary as a
defaultdict in order to use a generic function for logging stats.

Modify base agent rollout method to retrun the final state from the
rollout. Until now this method returned the states as a tensor of shape
(b, t, q) which means that the final state from each trajectory is not
returned. When using Monte-Carlo evaluation the final state is not
needed as we have no information for the return when starting from this
state. The method now returns a tensor of shape (b, t+1, q) which
includes the final state. This state is needed for bootstrapping
methods.