Add Actor-Critic agent
Add an implementation of an actor-critic agent. The agent is an online
advantage actor critic (A2C).
Although the logic of the algorithm allows for online updates, this is
going to be hard to implement. The reason for this is that the
environment operates on a batch of states and extends all trajectories
in parallel. Thus, at later stages of the trajectories, the batch of
states may consist of a smaller ammount of actual observations.
Another reason against online learning is that our current reward
function provides useful signals only at the end of each episode.