Bug fixes and minor changes
The script for training the AC agent was mistakenly using
args.batch_size isntead of args.env_batch.

Log a message when using pre-trained model for the PG agent.

Complete the docstring for the fcnn init method.

Fix a bug in the base agent `rollout` method. Iterating over a range
raises and exception `StopIteration` at the end of the loop, and thus
the value of the variable `i` does not increase outside the boundary
interval.