Bugfixes + Stochasticity + Renaming
Bugfix 1. Phase normalization used to yield states with imaginary
part of first vector component != 0 due to float division.
The fix is to explicitly set the imaginary component to 0 after
division.

Bugfix 2: The diagonalization of reduced density matrices in step()
function was suffering from precision  too - applying an action twice
used to change the state on the second time which is not
mathematically correct. The reason for this was
the small off-diagonal values in RDMs.
The fix for this was to clip values < 1e-7 to 0.

Stochasticity: The environment now has stochastic feature.
To enable it  pass "stochastic=True" in __init__()

Renaming: Functions in "util.py" have leading underscore removed.