Revise episode length calculation
During training we keep statistics of the number of steps in each
episode.

Until now we were calculating the episode length only for episodes for
which the agent managed to solve the environment. Note that in the
calculation we multiply by `~mask[:-1]`, i.e. if the last element is
masked then the episode has finished. It should be mentioned that this
calculation is **INCORRECT**. We should be using `done[:-1]` instead. In
case the agent needs exactly the same amount of steps as the episode
length, then we have `done[-1] = 1`, but `~mask[-1]=0` -- the last
element leads to done, but the last action should not be masked.

However, we should not be focusing only on solved trajectories because
this statistic is misleading. Instead we will be returning the number of
non-masked steps as is.