Add new obs_fn to the quantum env
The current obs_fn for 2-qubit rdms considers the different combinations
of qubits and creates the 2-qubit rmds for each combination. The new
function considers the reversed combinations as well. For each
combination (i,j) we create the rdm for qubits (i,j) and the rdm for
qubits (j,i) -- the reversed pair. The observation contains the average
of the two obtained rdms.